Adding Excel file managers
- Added excel file managers that extracts data from a single excel file
and modified selector and candidate in consequence. Now, we only gice
one path to selector which is a path to an excel file containing one
sheet for the candidates and another one for the courses. The
distribution is also written into the same excel file into a new sheet.
- Adding a CourseCandidate as to compare two candidates on a specific
course. The comparision is course dependant and it is not necessary to
change the candidates attributes to compare them with a course; with
this new class, a new object is created when comparing two candidates.
- Comparision test is made into this new class (lt, eq, ...) and fixed a
logic bug when comparing attributes. Now, there is a PRIORITY list
inside candidates that specifies the order of comparision when sorting a
list of candidates.
- Modified setup.py to include the openpyxl package that is needed to
read and manipulate excel files.